preparations for the junit5 module
- separation of the junit4 and junit5 code into different modules.
- core module to share code between different TDD frameworks
