+++ testing/junit5/pom.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <parent>
+    <artifactId>testing</artifactId>
+    <groupId>com.fnproject.fn</groupId>
+    <version>1.0.0-SNAPSHOT</version>
+  </parent>
+  <modelVersion>4.0.0</modelVersion>
+
+  <artifactId>testing-junit5</artifactId>
+
+  <dependencies>
+    <dependency>
+      <groupId>com.fnproject.fn</groupId>
+      <artifactId>testing-core</artifactId>
+    </dependency>
+  </dependencies>
+
+</project>
